Updates include: (SearchCard/*) - edit index.js to be the exporter and add SearchCard.js file   (/SearchCardMeta)     - update props and add the ‘size’ prop for label font size     - add the condition for displaying mixed platform icon for the platform ([experiment.js]) - use SearchCard/[SearchCardHeader, SearchCardCTAs, SearchCardMeta] components and pass necessary props (IconBadge) - (minor) rename ‘RNA’ to be ‘RanSeq’
