+++ src/components/SearchCard/SearchCardMeta.js
@@ -1,6 +1,8 @@
+/* eslint-disable no-nested-ternary */
 import { useResponsive } from 'hooks/useResponsive'
 import { formatNumbers } from 'helpers/formatNumbers'
 import { formatString } from 'helpers/formatString'
+import { isArray } from 'helpers/isArray'
 import { formatPlatformName } from 'helpers/formatPlatformName'
 import { Box } from 'grommet'
 import { IconBadge } from 'components/shared/IconBadge'
@@ -8,9 +10,14 @@ import { Row } from 'components/shared/Row'
 import { TextNull } from 'components/shared/TextNull'
 
 export const SearchCardMeta = ({
-  metaData: { downloadableSamples, organismNames, platformNames, technology }
+  downloadableSamples,
+  organismNames,
+  platformNames,
+  technology,
+  size = 'small'
 }) => {
   const { setResponsive } = useResponsive()
+  const technologyName = isArray(technology) ? technology.join('') : technology
 
   return (
     <Row
@@ -32,11 +39,9 @@ export const SearchCardMeta = ({
       <Box flex="grow" width={setResponsive('100%', '100%', { max: '30%' })}>
         {organismNames.length > 0 ? (
           <IconBadge
-            label={organismNames.map(
-              (organismName, i) =>
-                `${i > 0 ? ', ' : ''}${formatString(organismName)}`
-            )}
+            label={organismNames.map(formatString).join(', ')}
             name="Organism"
+            size={size}
           />
         ) : (
           <TextNull text="No species" />
@@ -46,16 +51,21 @@ export const SearchCardMeta = ({
         <IconBadge
           label={`${formatNumbers(downloadableSamples)} Downloadable Samples`}
           name="Samples"
+          size={size}
         />
       </Box>
       <Box flex="grow" width={setResponsive('100%', '100%', { max: '40%' })}>
         {platformNames.length > 0 ? (
           <IconBadge
-            label={platformNames.map(
-              (platformName, i) =>
-                `${i > 0 ? ', ' : ''}${formatPlatformName(platformName)}`
-            )}
-            name={technology === 'MICROARRAY' ? 'MicroArray' : 'Rna'}
+            label={platformNames.map(formatPlatformName).join(', ')}
+            name={
+              technologyName === 'MICROARRAY'
+                ? 'MicroArray'
+                : technologyName === 'RNA-SEQ'
+                ? 'RnaSeq'
+                : 'MixedPlatform'
+            }
+            size={size}
           />
         ) : (
           <TextNull text="No platform available" />
